refactor: move convenience fn
To slightly more sensical place.
